refactor(core): hoist page-script shared runner into core [COU-1]
